requirements.txt
openai
langgraph
cadquery
pythonocc-core
